api/pyproject.toml
[tool.poetry]
name = "granite-guardian"
version = "0.1.0"
description = ""
authors = ["Yash Sawlani <yash.sawlani@ibm.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12, <3.13"
ibm-watsonx-ai = "^1.1.25"
langchain = "^0.3.11"
python-dotenv = "^1.0.1"
fastapi = "^0.115.6"
uvicorn = "^0.32.1"
sendgrid = "^6.11.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
